move emails to their own file
